pyproject.toml
[tool.poetry]
name = "library-managment"
version = "0.1.0"
description = ""
authors = ["Mahdi Ghiasi <mahdiw.ghiasi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2.13"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
drf-spectacular = "^0.22.1"
dj-rest-auth = "^2.2.4"
djangorestframework-simplejwt = "^5.2.0"
drf-nested-routers = "^0.93.4"
redis = "^4.3.3"
celery = "^5.2.7"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
django-debug-toolbar = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
